# apidog.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EvenIt Backend API",
    "version": "1.0.0",
    "description": "Role-based event marketplace backend for auth, onboarding, listings, bookings, subscriptions, uploads, and admin moderation."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "VenueEntity": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "information": {
            "type": "object"
          },
          "pricing": {
            "type": "object"
          },
          "capacity": {
            "type": "object"
          },
          "media": {
            "type": "object"
          },
          "availability": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          },
          "publishStatus": {
            "type": "string",
            "enum": [
              "pending",
              "published",
              "rejected"
            ]
          },
          "approvedBy": {
            "nullable": true,
            "type": "object"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reviews": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "VenueListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              },
              "hasNextPage": {
                "type": "boolean"
              },
              "hasPrevPage": {
                "type": "boolean"
              }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueEntity"
            }
          }
        }
      },
      "VenueResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/VenueEntity"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/venue-provider/my-venues": {
      "get": {
        "tags": [
          "VenueProvider"
        ],
        "summary": "Get own venues",
        "description": "Authenticated venue providers can view only their own venues and optionally filter by publish status such as pending or published.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "in": "query",
            "name": "publishStatus",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "published",
                "rejected"
              ]
            },
            "description": "Use `pending` to see requested venues awaiting approval, `published` to see approved venues, or omit it to see both."
          }
        ],
        "responses": {
          "200": {
            "description": "Own venue list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/venue-provider/venues": {
      "get": {
        "tags": [
          "VenueProvider"
        ],
        "summary": "Get published venues (public)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated published venues",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "VenueProvider"
        ],
        "summary": "Create a venue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Submit multipart form-data. Put the venue JSON in the `payload` field and attach optional image files in `images` or `image`. The backend uploads those files and appends the resulting URLs to `media.galleryImages`.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "payload"
                ],
                "properties": {
                  "payload": {
                    "type": "string",
                    "description": "JSON string matching VenueCreateRequest. Include `media.galleryImages` only for already-hosted image URLs. Files sent in `images` or `image` are uploaded by the backend automatically.",
                    "example": "{\"information\":{\"venueName\":\"Grand Hall\",\"venueType\":\"Banquet\",\"description\":\"Premium event venue in central Dhaka.\",\"addressLine\":\"123 Main Road\",\"city\":\"Dhaka\",\"area\":\"Farmgate\"},\"pricing\":{\"basePrice\":5000,\"currency\":\"BDT\",\"discount\":{\"type\":\"percentage\",\"value\":10},\"amenities\":{\"wifi\":true,\"parking\":true,\"ac\":true}},\"capacity\":{\"maximumGuests\":300},\"media\":{\"galleryImages\":[],\"videoUrl\":\"https://www.youtube.com/watch?v=abc123\"},\"availabilityCalendar\":[{\"date\":\"2026-03-20\",\"hours\":[8,9,10]}]}"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Optional venue gallery images. Up to 10 files."
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional single-image field supported for client compatibility."
                  }
                }
              },
              "examples": {
                "validVenue": {
                  "summary": "Valid venue creation payload",
                  "value": {
                    "payload": "{\"information\":{\"venueName\":\"Grand Hall\",\"venueType\":\"Banquet\",\"description\":\"Premium event venue in central Dhaka.\",\"addressLine\":\"123 Main Road\",\"city\":\"Dhaka\",\"area\":\"Farmgate\"},\"pricing\":{\"basePrice\":5000,\"currency\":\"BDT\",\"discount\":{\"type\":\"percentage\",\"value\":10},\"amenities\":{\"wifi\":true,\"parking\":true,\"ac\":true}},\"capacity\":{\"maximumGuests\":300},\"media\":{\"galleryImages\":[],\"videoUrl\":\"https://www.youtube.com/watch?v=abc123\"},\"availabilityCalendar\":[{\"date\":\"2026-03-20\",\"hours\":[8,9,10]}]}"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Venue created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/venue-provider/venues/{venueId}": {
      "get": {
        "tags": [
          "VenueProvider"
        ],
        "summary": "Get one published venue by id (public)",
        "parameters": [
          {
            "in": "path",
            "name": "venueId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Venue details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/VenueEntity"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Venue not found"
          }
        }
      },
      "patch": {
        "tags": [
          "VenueProvider"
        ],
        "summary": "Update venue by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Submit multipart form-data. Put the partial venue JSON in the `payload` field and attach optional image files in `images` or `image`. The backend uploads those files and appends the resulting URLs to `media.galleryImages`.",
        "parameters": [
          {
            "in": "path",
            "name": "venueId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "payload"
                ],
                "properties": {
                  "payload": {
                    "type": "string",
                    "description": "JSON string matching VenueUpdateRequest. Files sent in `images` or `image` are uploaded by the backend automatically.",
                    "example": "{\"capacity\":{\"maximumGuests\":550},\"media\":{\"galleryImages\":[],\"videoUrl\":\"https://youtube.com/watch?v=updated-venue\"},\"availabilityCalendar\":[{\"date\":\"2026-04-19\",\"hours\":[18]}]}"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Optional venue gallery images. Up to 10 files."
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional single-image field supported for client compatibility."
                  }
                }
              },
              "examples": {
                "validVenueUpdate": {
                  "summary": "Valid venue update payload",
                  "value": {
                    "payload": "{\"capacity\":{\"maximumGuests\":550},\"media\":{\"galleryImages\":[],\"videoUrl\":\"https://youtube.com/watch?v=updated-venue\"},\"availabilityCalendar\":[{\"date\":\"2026-04-19\",\"hours\":[18]}]}"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Venue updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Venue not found"
          }
        }
      },
      "delete": {
        "tags": [
          "VenueProvider"
        ],
        "summary": "Delete venue by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "venueId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Venue deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Venue deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Venue not found"
          }
        }
      }
    },
...
  },
...
}